requirements.txt
greenlet==3.0.3
MouseInfo==0.1.3
playwright==1.47.0
psycopg2==2.9.10
PyAutoGUI==0.9.54
pyee==12.0.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
pytweening==1.2.0
selectolax==0.3.21
SQLAlchemy==2.0.36
typing_extensions==4.12.2
